I had need of dynamically changing the text color of an UILabel and I find dealing with NSAttributedText to be clumsy. This commit allows the "textColor" attribute of the UILabel to be modified in the same way as the "backgroundColor" attribute of the UIView.
I added the appropriate code and tests, following the style of the already written code. Also modified the Readme file to include "dynTextColor" in the table which displays the available dyn* variables

Please go easy on me, this is my first commit on an open source project. I am intentionally making it a small change, I may develop more in the future.