Fix Retain Cycle in Swift `KeyPath` KVO
Using the new KVO signal that uses Swift’s `KeyPath`, there was a retain cyle that was stopping the base object from being deallocated. Additionally, it was not firing the completed event when the base object deallocated (once the retain cycle was fixed).

The newly added test is verified to be vaild. It failed prior to the changes.